Merges unit tests into the controllers directory
